+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.4 - 2025-08-27]
+## Unreleased changes
+### Added
+- ignore_descreasing_sequences DSL method
+### Fixed
+- DecreasingSequenceError error message description fixed
+
+## [1.0.5 - 2025-08-27]
 ### Added
 - DecreasingSequenceError provides diagnostic info